+++ src/main/java/com/yunche/loan/domain/vo/InfoSupplementVO.java
@@ -46,7 +46,7 @@ public class InfoSupplementVO {
     /**
      * 增补类型：  1-电审资料增补;  2-资料审核增补;
      */
-    private Integer supplementType;
+    private Byte supplementType;
     /**
      * 要求增补内容
      */
@@ -80,6 +80,7 @@ public class InfoSupplementVO {
      * 担保人列表
      */
 //    private List<CustomerFile> guarantorList;
+
     /**
      * 紧急联系人列表
      */
